a5teroids: Suppress warning about double to int conversion.
cherry-pick https://alleg.svn.sourceforge.net/svnroot/alleg/allegro/branches/5.1@13950

From: Peter Wang <tjaden@example.org>

+++ demos/a5teroids/src/UFO.cpp
@@ -11,7 +11,7 @@ bool UFO::logic(int step)
       return false;
    }
 
-   int now = al_get_time() * 1000;
+   int now = (int) (al_get_time() * 1000.0);
    if (now > nextShot) {
       nextShot = now + SHOT_SPEED;
       ResourceManager& rm = ResourceManager::getInstance();
